yash project/requirements.txt
streamlit==1.42.2
pandas==2.2.3
numpy==1.26.4
plotly==5.24.1
matplotlib==3.9.2
seaborn==0.13.2
python-dotenv==1.0.1
groq==0.13.1 